New relationship task user

+++ back-end/controller/task.routes.ts
@@ -1,5 +1,6 @@
 import taskService from "../service/task.service";
-import express, { Request, Response } from 'express';
+import express, { NextFunction, Request, Response } from 'express';
+import { TaskInput } from "../types";
 
 const taskRouter = express.Router();
 
@@ -19,8 +20,12 @@ const taskRouter = express.Router();
  *         deadline:
  *           type: string
  *           format: date-time
+ *         completed:
+ *           type: boolean
  *         project:
  *           $ref: "#/components/schemas/Project"
+ *         user:
+ *           $ref: "#/components/schemas/User"
  *       required:
  *         - name
  *         - description
@@ -29,8 +34,18 @@ const taskRouter = express.Router();
  *     TaskInput:
  *       type: object
  *       properties:
- *         id:
+ *         name:
+ *           type: string
+ *         description:
+ *           type: string
+ *         deadline:
+ *           type: string
+ *           format: date-time
+ *         projectId:
  *           type: integer
+ *     TaskInputUpdate:
+ *       type: object
+ *       properties:
  *         name:
  *           type: string
  *         description:
@@ -40,8 +55,6 @@ const taskRouter = express.Router();
  *           format: date-time
  *         completed:
  *           type: boolean
- *         project:
- *           $ref: "#/components/schemas/ProjectInput"
  */
 
 
@@ -74,13 +87,15 @@ const taskRouter = express.Router();
  *                   type: string
  */
 
-taskRouter.get('/', async (req: Request, res: Response) => {
+taskRouter.get('/', async (req: Request & {auth: any}, res: Response, next: NextFunction) => {
     try {
-        const tasks = await taskService.getAllTasks();
+        const currentUser = req.auth.id;
+        const role = req.auth.role;
+        const tasks = await taskService.getAllTasks({currentUser, role});
         res.status(200).json(tasks);
     }
     catch (error) {
-        res.status(400).json({ status: 'error', errorMessage: error.message });
+        next(error);
     }
 });
 
@@ -116,13 +131,15 @@ taskRouter.get('/', async (req: Request, res: Response) => {
  *                   type: string
  */
 
-taskRouter.get('/:id', async (req: Request, res: Response) => {
+taskRouter.get('/:id', async (req: Request & {auth: any}, res: Response, next: NextFunction) => {
     try {
-        const task = await taskService.getTaskById(parseInt(req.params.id));
+        const role = req.auth.role;
+        const currentUser = req.auth.id;
+        const task = await taskService.getTaskById({id: parseInt(req.params.id), currentUser, role});
         res.status(200).json(task);
     }
     catch (error) {
-        res.status(400).json({ status: 'error', errorMessage: error.message });
+        next(error);
     }
 });
 
@@ -160,13 +177,15 @@ taskRouter.get('/:id', async (req: Request, res: Response) => {
  *                   type: string
  */
 
-taskRouter.get('/project/:id', async (req: Request, res: Response) => {
+taskRouter.get('/project/:id', async (req: Request & {auth: any}, res: Response, next: NextFunction) => {
     try {
-        const tasks = await taskService.getTaskByProjectId(parseInt(req.params.id));
+        const role = req.auth.role;
+        const currentUser = req.auth.id;
+        const tasks = await taskService.getTasksByProjectId({id: parseInt(req.params.id), currentUser, role});
         res.status(200).json(tasks);
     }
     catch (error) {
-        res.status(400).json({ status: 'error', errorMessage: error.message });
+        next(error);
     }
 });
 
@@ -182,7 +201,7 @@ taskRouter.get('/project/:id', async (req: Request, res: Response) => {
  *           schema:
  *             $ref: '#/components/schemas/TaskInput'
  *     responses:
- *       201:
+ *       200:
  *         description: Successful response with the created task.
  *         content:
  *           application/json:
@@ -201,31 +220,39 @@ taskRouter.get('/project/:id', async (req: Request, res: Response) => {
  *                   type: string
  */
 
-taskRouter.post('/', async (req: Request, res: Response) => {
+taskRouter.post('/', async (req: Request & {auth: any}, res: Response, next: NextFunction) => {
     try {
-        // console.log(req.body)
-        const task = await taskService.createTask(req.body);
-        res.status(201).json(task);
+        const task = <TaskInput>req.body;
+        const currentUser = req.auth.id;
+        const response = await taskService.createTask({task, currentUser});
+        res.status(200).json(response);
     }
     catch (error) {
-        res.status(400).json({ status: 'error', errorMessage: error.message });
+        next(error);
     }
 });
 
 /**
  * @swagger
- * /tasks:
+ * /tasks/{id}:
  *   put:
- *     summary: Create a new task.
+ *     summary: Update a task.
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         description: ID of the task to update.
+ *         schema:
+ *           type: integer
  *     requestBody:
  *       required: true
  *       content:
  *         application/json:
  *           schema:
- *             $ref: '#/components/schemas/TaskInput'
+ *             $ref: '#/components/schemas/TaskInputUpdate'
  *     responses:
- *       201:
- *         description: Successful response with the created task.
+ *       200:
+ *         description: Successful response with the updated task.
  *         content:
  *           application/json:
  *             schema:
@@ -243,13 +270,58 @@ taskRouter.post('/', async (req: Request, res: Response) => {
  *                   type: string
  */
 
-taskRouter.put('/:id', async (req: Request, res: Response) => {
+taskRouter.put('/:id', async (req: Request & {auth: any}, res: Response, next: NextFunction) => {
     try {
-        const task = await taskService.updateTask(req.body);
-        res.status(201).json(task);
+        const currentUser = req.auth.id;
+        const role = req.auth.role;
+        const task = <TaskInput>req.body;
+        const response = await taskService.updateTask({targetTaskId: parseInt(req.params.id), updatedInfo: task, currentUser, role});
+        res.status(200).json(response);
     }
     catch (error) {
-        res.status(400).json({ status: 'error', errorMessage: error.message });
+        next(error);
+    }
+});
+
+/**
+ * @swagger
+ * /tasks/completeTask/{id}:
+ *   put:
+ *     summary: Complete a task.
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         description: ID of the task to complete.
+ *         schema:
+ *           type: integer
+ *     responses:
+ *       200:
+ *         description: Successful response with the completed task.
+ *         content:
+ *           application/json:
+ *             schema:
+ *               $ref: '#/components/schemas/Task'
+ *       400:
+ *         description: Bad request with an error message.
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 status:
+ *                   type: string
+ *                 errorMessage:
+ *                   type: string
+ */
+taskRouter.put('/completeTask/:id', async (req: Request & {auth:any}, res: Response, next: NextFunction) => {
+    try {
+        const currentUser = req.auth.id;
+        const response = await taskService.completeTask({id: parseInt(req.params.id), currentUser});
+        res.status(200).json(response);
+    }
+    catch (error) {
+        next(error);
     }
 });
 
@@ -285,13 +357,15 @@ taskRouter.put('/:id', async (req: Request, res: Response) => {
  *                   type: string
  */
 
-taskRouter.delete('/:id', async (req: Request, res: Response) => {
+taskRouter.delete('/:id', async (req: Request & {auth: any}, res: Response, next: NextFunction) => {
     try {
-        const task = await taskService.deleteById(parseInt(req.params.id));
+        const currentUser = req.auth.id;
+        const role = req.auth.role;
+        const task = await taskService.deleteById({id: parseInt(req.params.id), currentUser, role});
         res.status(200).json(task);
     }
     catch (error) {
-        res.status(400).json({ status: 'error', errorMessage: error.message });
+        next(error);
     }
 });
